backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-jose
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
redis
minio
email-validator
python-jose[cryptography]
